==> r_scripts/04_basic_recipe_tree.R <==
# STAT 301-2 Final Project: Predicting Social Media Shares ----
# Tree Recipe

# load packages ----
library(tidyverse)
library(tidymodels)
library(here)

# handle common conflicts ----
tidymodels_prefer()

# set seed ----
set.seed(3012)

# load files ----
load(here("data_splits/articles_train.rda"))

# recipe ----
basic_recipe_tree <- recipe(shares_log ~., 
                            data = articles_train) |>
  step_rm(url, timedelta, shares) |>
  step_zv(all_predictors()) |>
  step_dummy(all_nominal_predictors(), one_hot = TRUE) |>
  step_normalize(all_numeric_predictors())

prep(basic_recipe_tree) |>
  bake(new_data = NULL) |>
  glimpse()

save(basic_recipe_tree, file = "recipes/basic_recipe_tree.rda")
